requirements.txt
streamlit==1.24.0
pandas==2.0.3
numpy==1.24.3
plotly==5.15.0
requests==2.31.0
scipy>=1.9.0
scikit-learn>=1.0.0
python-dotenv>=0.19.0
openpyxl>=3.0.0
watchdog
rdkit>=2023.3.1
dataclasses>=0.6
typing-extensions>=4.5.0
matplotlib>=3.7.1
